feat: consolidate mod tools and extract toggle persist watcher
