+++ src/shared/types.gen.ts
@@ -87,16 +87,16 @@ export type IpcHandlers = {
     "drive:get:item": (...args: Parameters<typeof desktop.service.drive.get.item>) => ReturnType<typeof desktop.service.drive.get.item>;
     "drive:patch:rename": (...args: Parameters<typeof desktop.service.drive.patch.rename>) => ReturnType<typeof desktop.service.drive.patch.rename>;
     "drive:post:dir": (...args: Parameters<typeof desktop.service.drive.post.dir>) => ReturnType<typeof desktop.service.drive.post.dir>;
-    "ftm:cancelRun": (...args: Parameters<typeof desktop.service.fixTools.cancelRun>) => ReturnType<typeof desktop.service.fixTools.cancelRun>;
-    "ftm:createPreset": (...args: Parameters<typeof desktop.service.fixTools.createPreset>) => ReturnType<typeof desktop.service.fixTools.createPreset>;
-    "ftm:deletePreset": (...args: Parameters<typeof desktop.service.fixTools.deletePreset>) => ReturnType<typeof desktop.service.fixTools.deletePreset>;
-    "ftm:deleteScript": (...args: Parameters<typeof desktop.service.fixTools.deleteScript>) => ReturnType<typeof desktop.service.fixTools.deleteScript>;
-    "ftm:getPresets": (...args: Parameters<typeof desktop.service.fixTools.getPresets>) => ReturnType<typeof desktop.service.fixTools.getPresets>;
-    "ftm:getScripts": (...args: Parameters<typeof desktop.service.fixTools.getScripts>) => ReturnType<typeof desktop.service.fixTools.getScripts>;
-    "ftm:runPreset": (...args: Parameters<typeof desktop.service.fixTools.runPreset>) => ReturnType<typeof desktop.service.fixTools.runPreset>;
-    "ftm:runScript": (...args: Parameters<typeof desktop.service.fixTools.runScript>) => ReturnType<typeof desktop.service.fixTools.runScript>;
-    "ftm:saveScript": (...args: Parameters<typeof desktop.service.fixTools.saveScript>) => ReturnType<typeof desktop.service.fixTools.saveScript>;
-    "ftm:sendInput": (...args: Parameters<typeof desktop.service.fixTools.sendInput>) => ReturnType<typeof desktop.service.fixTools.sendInput>;
+    "ftm:cancelRun": (...args: Parameters<typeof desktop.service.modTools.fixTool.cancelRun>) => ReturnType<typeof desktop.service.modTools.fixTool.cancelRun>;
+    "ftm:createPreset": (...args: Parameters<typeof desktop.service.modTools.fixTool.createPreset>) => ReturnType<typeof desktop.service.modTools.fixTool.createPreset>;
+    "ftm:deletePreset": (...args: Parameters<typeof desktop.service.modTools.fixTool.deletePreset>) => ReturnType<typeof desktop.service.modTools.fixTool.deletePreset>;
+    "ftm:deleteScript": (...args: Parameters<typeof desktop.service.modTools.fixTool.deleteScript>) => ReturnType<typeof desktop.service.modTools.fixTool.deleteScript>;
+    "ftm:getPresets": (...args: Parameters<typeof desktop.service.modTools.fixTool.getPresets>) => ReturnType<typeof desktop.service.modTools.fixTool.getPresets>;
+    "ftm:getScripts": (...args: Parameters<typeof desktop.service.modTools.fixTool.getScripts>) => ReturnType<typeof desktop.service.modTools.fixTool.getScripts>;
+    "ftm:runPreset": (...args: Parameters<typeof desktop.service.modTools.fixTool.runPreset>) => ReturnType<typeof desktop.service.modTools.fixTool.runPreset>;
+    "ftm:runScript": (...args: Parameters<typeof desktop.service.modTools.fixTool.runScript>) => ReturnType<typeof desktop.service.modTools.fixTool.runScript>;
+    "ftm:saveScript": (...args: Parameters<typeof desktop.service.modTools.fixTool.saveScript>) => ReturnType<typeof desktop.service.modTools.fixTool.saveScript>;
+    "ftm:sendInput": (...args: Parameters<typeof desktop.service.modTools.fixTool.sendInput>) => ReturnType<typeof desktop.service.modTools.fixTool.sendInput>;
     "mod:addGame": (...args: Parameters<typeof desktop.service.mod.fn.addGame>) => ReturnType<typeof desktop.service.mod.fn.addGame>;
     "mod:applyPreset": (...args: Parameters<typeof desktop.service.mod.fn.applyPreset>) => ReturnType<typeof desktop.service.mod.fn.applyPreset>;
     "mod:copyFolder": (...args: any[]) => any;
@@ -164,9 +164,9 @@ export type IpcHandlers = {
     "setting:net:setProxy": (...args: Parameters<typeof desktop.setting.net.setProxy>) => ReturnType<typeof desktop.setting.net.setProxy>;
     "setting:xxmi:getPersistToggles": (...args: Parameters<typeof desktop.setting.xxmi.getPersistToggles>) => ReturnType<typeof desktop.setting.xxmi.getPersistToggles>;
     "setting:xxmi:setPersistToggles": (...args: any[]) => any;
-    "tools:buildNewD3DDLL": (...args: Parameters<typeof desktop.service.tools.buildNewD3DDLL>) => ReturnType<typeof desktop.service.tools.buildNewD3DDLL>;
-    "tools:getGIMIPath": (...args: Parameters<typeof desktop.service.tools.getGIMIPath>) => ReturnType<typeof desktop.service.tools.getGIMIPath>;
-    "tools:saveGIMIPath": (...args: Parameters<typeof desktop.service.tools.saveGIMIPath>) => ReturnType<typeof desktop.service.tools.saveGIMIPath>;
+    "tools:buildNewD3DDLL": (...args: Parameters<typeof desktop.service.modTools.dllBuilder.buildNewD3DDLL>) => ReturnType<typeof desktop.service.modTools.dllBuilder.buildNewD3DDLL>;
+    "tools:getGIMIPath": (...args: Parameters<typeof desktop.service.modTools.dllBuilder.getGIMIPath>) => ReturnType<typeof desktop.service.modTools.dllBuilder.getGIMIPath>;
+    "tools:saveGIMIPath": (...args: Parameters<typeof desktop.service.modTools.dllBuilder.saveGIMIPath>) => ReturnType<typeof desktop.service.modTools.dllBuilder.saveGIMIPath>;
     "transfer:cancel": (...args: Parameters<typeof desktop.service.transfer.cancelTransfer>) => ReturnType<typeof desktop.service.transfer.cancelTransfer>;
     "transfer:clear": (...args: any[]) => any;
     "transfer:list": (...args: Parameters<typeof desktop.service.transfer.getAllTransfer>) => ReturnType<typeof desktop.service.transfer.getAllTransfer>;
